refactor: move Submission queries to subdirectory
