+++ clients/web/src/lib/api/queries/index.ts
@@ -1,9 +1,8 @@
 export * from "./GamesAndModes.store";
-export * from "./SubmissionStats.store";
 export * from "./RestrictedPlayers.store";
 export * from "./GetMemberByUserId";
-export * from "./Submission.store";
 export * from "./CurrentUser.store";
 
 export * from "./league";
 export * from "./scrims";
+export * from "./submissions";
